Fix bug & move data
