+++ firefox-ios/Client/Frontend/Browser/MainMenu/Redux/MainMenuState.swift
@@ -3,11 +3,13 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import Common
+import MenuKit
 import Shared
 import Redux
 
 struct MainMenuState: ScreenState, Equatable {
     var windowUUID: WindowUUID
+    var menuElements: [[MenuElement]]
     var shouldDismiss: Bool
 
     init(appState: AppState, uuid: WindowUUID) {
@@ -22,6 +24,7 @@ struct MainMenuState: ScreenState, Equatable {
 
         self.init(
             windowUUID: mainMenuState.windowUUID,
+            menuElements: mainMenuState.menuElements,
             shouldDismiss: mainMenuState.shouldDismiss
         )
     }
@@ -29,32 +32,73 @@ struct MainMenuState: ScreenState, Equatable {
     init(windowUUID: WindowUUID) {
         self.init(
             windowUUID: windowUUID,
+            menuElements: [],
             shouldDismiss: false
         )
     }
 
     private init(
         windowUUID: WindowUUID,
+        menuElements: [[MenuElement]],
         shouldDismiss: Bool
     ) {
         self.windowUUID = windowUUID
         self.shouldDismiss = shouldDismiss
+        self.menuElements = MainMenuConfigurationUtility().populateMenuElements(with: windowUUID)
     }
 
     static let reducer: Reducer<Self> = { state, action in
         guard action.windowUUID == .unavailable || action.windowUUID == state.windowUUID else { return state }
 
         switch action.actionType {
+        case MainMenuActionType.viewDidLoad:
+            return MainMenuState(
+                windowUUID: state.windowUUID,
+                menuElements: state.menuElements,
+                shouldDismiss: false
+            )
         case MainMenuActionType.closeMenu:
             return MainMenuState(
                 windowUUID: state.windowUUID,
+                menuElements: state.menuElements,
                 shouldDismiss: true
             )
         default:
             return MainMenuState(
                 windowUUID: state.windowUUID,
+                menuElements: state.menuElements,
                 shouldDismiss: false
             )
         }
+    }
+}
+
+struct MainMenuConfigurationUtility {
+    func populateMenuElements(with uuid: WindowUUID) -> [[MenuElement]] {
+        let fakeMenuItem = MenuElement(
+            title: "Test title",
+            iconName: "",
+            isEnabled: true,
+            isActive: false,
+            a11yLabel: "",
+            a11yHint: "",
+            a11yId: "",
+            action: {
+                store.dispatch(
+                    MainMenuAction(
+                        windowUUID: uuid,
+                        actionType: MainMenuActionType.closeMenu
+                    )
+                )
+            }
+        )
+
+        return [
+            [fakeMenuItem, fakeMenuItem],
+            [fakeMenuItem, fakeMenuItem, fakeMenuItem, fakeMenuItem, fakeMenuItem],
+            [fakeMenuItem, fakeMenuItem, fakeMenuItem, fakeMenuItem, fakeMenuItem],
+            [fakeMenuItem, fakeMenuItem, fakeMenuItem, fakeMenuItem, fakeMenuItem],
+            [fakeMenuItem, fakeMenuItem],
+        ]
     }
 }
